backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.28
pydantic==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
aiofiles==24.1.0
psycopg2-binary==2.9.9
redis==5.0.1 